Add a simple sir-trevor widget to include an image and title for a document.
This widget currently will just take a known ID of a document in the local
index and display the configured title and image thumbnail.

Their is a checkbox to configure weather to show the title or not in
case the user just wants the image for a document that has a title.