Add UniversalViewer
This commit adds a `package.json` so that
is possible to install `universalviewer`
with yarn in Rails 5.1+ applications.

To install uv in your development
enviroment run `yarn run install-uv`.

In production there should be a capistrano task that
moves the uv assets to the `/public` folder.

In the catalog controller for the spotlight app
the `universal_viewer_default` partial should be used
instead of `openseadragon`. This
uses the `<iframe>` version of UniversalViewer. It's
done this way because including UV directly is not
compatible with jQuery almond and other JS included
in `spotlight.js`.

Documentation on the options available for the viewer
is here:

https://github.com/UniversalViewer/universalviewer/wiki/Embedding#hash-parameters

There is a `ManifestService` and a `#iiif_manifest` helper that are
used in views.