Add a helper for getting a IIIF manifest URL
This commit adds a service and helper for retrieving the IIIF manifest
from a SolrDocument. Currently trying to access this URL
for an item on a view is kind of awkward:

```
document['response']['response']['docs'][0]['iiif_manifest_url_ssi']
```

With this service and helper, you can access it in views by doing:

```
iiif_manifest(manifest_service: Spotlight::ManifestService.new(document: document))
```

The functionality of the service could be expanded to additional
checking and other useful things in the future.