Pass format explicitly through page form/controller when updating.
The polymorphic routing seems to choke on allowing us to handle both the has_many feature/about pages relation and the has_one home_page relation the same.

In the case of the has_one relationship we're seeing the "slug" or ID of the page being appended like a format to the URL.

If we explicitly pass a nil format through the form (which should always be HTML and is handled appropriately as the default) and then pass the format param explicitly back through the redirect after update (so updating via API will still work in theory) and things appear to work out both on the home and feature/about pages side of things.