Force a Leaflet layer to have a container when we load a layer w/o one.
This is typically in the case that the Layer failed to be set by Leaflet-iiif (e.g. the info.json fails to load for one reason or another).

This allows users to replace the image in the viewer successfully and swap it out w/ a new on in this case.