generate iiif manifest URLs from the resource ID, not the image ID
The manifest controller expects the URL to include the exhibit ID and resource
ID, but the indexer has been using the ID of the resource's FeaturedImage
instead.  This doesn't always cause issues, because in many situations the ID of
a resource's FeaturedImage and the ID of the resource itself have the same
numerical representation.  But those numbers can fall out of sync, and manifests
will fail to load.