Replace typeahead with auto-complete-element
Replaces twitter typeahead fields, with the exception of the exhibit tag
field, with GitHub auto-complete-element. Fully removing the typeahead
library and SCSS should happen when the replacement for the exhibit tag
is merged. Twitter typeahead has been out of support for years and
auto-complete-element is in use by Blacklight currently.