Try a temporary BS4 workaround
Somewhere we lost the jQuery integration for BS4 (e.g., modal, tab) in some contexts (e.g., zpr_links.js) when importing BS4 in the app's application.js to be bundled with the default esbuild. It seems to work when included in the layout. Let's see if this passes the test suite.