Move to Propshaft, stop supporting Sprockets and Blacklight 7 (#3161)
* Fix bundle install step in generator

* Fix deprecation warning for expected boolean value

* Try versions that have corresponding NPM packages

The selection prompt may be causing problems on CI?

* Build the assets once during app generation

* Test using a newer version of Bootstrap 4

* Adjust Dockerfile settings for propshaft generator

* Handle missing blacklight frontend versions

* Use 'yarn add file' when running in CI

* Skip javascript on template installs

* Add turbo event to form observer

* Switch install generator to propshaft

* Try a temporary BS4 workaround

Somewhere we lost the jQuery integration for BS4 (e.g., modal, tab) in some contexts (e.g., zpr_links.js) when importing BS4 in the app's application.js to be bundled with the default esbuild. It seems to work when included in the layout. Let's see if this passes the test suite.

* Quote bootstrap versions in CI config

* Remove local copy of CheckboxSubmit

If we force at least Blacklight 7.38 and use the ESM module, Blacklight
handles this.

* Please let this stop the test from flapping

* Test on any version of Blacklight 8

Currently BL8 has corresponding frontend packages for all versions.

* Add packages individually

* Revert CI version changes

They are unnecessary. The Blacklight frontend package is based on the
actual installed gem version. After the switch to using '^', the highest
allowed version of bootstrap will be installed anyway.

* Require version of OSD rails that supports turbo

* Remove --skip-javascript workaround

* Precompile assets before testing

I think this is better suited here rather than in the generator.

* fix download_template_exhibit_bulk_updates

* fix loading bar on translation download

* Sync default Bootstrap versions

* Set data-turbo: false on download template link

* Use turbo streams for translation language removal

* Remove Bootstrap 4 include from layout

* Remove bundle install step from generator

Temp

* add css for btn blacklight icons

* Remove active support for Blacklight 7

* Add importmap generator

* Default to importmap in the Dockerfile

* Consolidate javascript

* Use the preserve-symlinks option in dev

* Remove unnecessary import

We are keeping the global SirTrevor object so we can walk this change back.

* Use best guess for spotlight-frontend package

* Remove unnecessary reference to window

* Revert unnecessary change

This change isn't needed when we are depending on global jQuery.

* Update the generator descriptions

* Add the spotlight package

---------

Co-authored-by: Cory Lown <corylown@stanford.edu>
Co-authored-by: dnoneill <noneill@stanford.edu>
Co-authored-by: Justin Coyne <jcoyne@justincoyne.com>
Co-authored-by: Chris Beer <cabeer@stanford.edu>